Adding Before and After Capability

+++ timelapse.py
@@ -25,7 +25,8 @@ def GoToPlace(inputPlace):
     helper.LocateAndClick('./common/closeSearch.png', helper.SMALL_PAUSE)
     
 
-
+def GetMode(startVal, endVal):
+    return (-1, 1)[startVal > endVal]
 def record(inputJSON):
     helper.PauseForEffect(helper.SMALL_PAUSE)
     if "place" in inputJSON:
@@ -37,7 +38,6 @@ def record(inputJSON):
 
     assert "start_count" in inputJSON, "Need to Specify start count"
     assert "end_count" in inputJSON, "Need to Specify end count"
-    assert "mode" in inputJSON, "Need to specify the mode of the timelapse (FORWARD or BACKWARD)"
     assert "image_type" in inputJSON, "Need to Specify Image type for snapshot"
     
     # Locate Time Lapse
@@ -47,31 +47,46 @@ def record(inputJSON):
         helper.ClickAndWait(x, y, helper.SMALL_PAUSE)
         startVal = inputJSON["start_count"]
         endVal = inputJSON["end_count"]
-        steps = startVal - endVal
-        # When it starts it starts at the Start
-        # So if we need forward we need to hit back
-        if inputJSON["mode"] == "FORWARD":
-            for i in range(0, steps + 1): # Need Extra as we are taking a snapshot after click
-                helper.ClickAndWait(x + backwardDelta[0], y + backwardDelta[1])
+        assert startVal != endVal, "Start and End Cannot be the same"
+        for i in range(0, startVal + GetMode(startVal, endVal)): # Need Extra as we are taking a snapshot after click
+            helper.ClickAndWait(x + backwardDelta[0], y + backwardDelta[1])
                 
+        helper.PauseForEffect(helper.MEDIUM_PAUSE)
+        print("Current Mode is {}".format(str(GetMode(startVal, endVal))))
 
         # Preprocessing before taking snapshot
         stepX = x
         stepY = y
-        if inputJSON["mode"] == "BACKWARD":
+        if GetMode(startVal, endVal) == -1:
             stepX = x + backwardDelta[0]
             stepY = y + backwardDelta[1]
-        if inputJSON["mode"] == "FORWARD":
+        if GetMode(startVal, endVal) == 1:
             stepX = x + forwardDelta[0]
             stepY = y + forwardDelta[1]
-        totalPad = int(math.ceil(math.log10(steps)))
-        currentImageType = helper.imageTypeMap[helper.ImageType[inputJSON["image_type"]]]
+
+        listToCapture = []
+        # Capture All Steps
+        if inputJSON["type"] == "TIMELAPSE":
+            listToCapture = [i in range(startVal, endVal + GetMode(startVal, endVal), GetMode(startVal, endVal))]
+        # Capture Only Start and End
+        if inputJSON["type"] == "BNA":
+            listToCapture.append(startVal)
+            listToCapture.append(endVal)
+        totalPad = int(math.ceil(math.log10(max(startVal, endVal))))
+        currentImageType = helper.imageTypeMap[helper.ImageType[inputJSON["image_type"]]]        
+        print(listToCapture)
+        
         
         
         # Click and Create image
-        for i in range(0, steps):
+        for i in range(startVal, endVal + (-1) * GetMode(startVal, endVal), (-1) * GetMode(startVal, endVal)):
             # Click
+            print("Current Step is {}".format(str(i)))
             helper.ClickAndWait(stepX, stepY, helper.SMALL_PAUSE)
+            # Skip the ones that are not necessary
+            if i not in listToCapture:
+                continue
+            print("Snapshot Step is {}".format(str(i)))
             # Create image
             pyautogui.screenshot(TIMELAPSE_IMAGE_NAME + str(i).zfill(totalPad) + currentImageType, region = imageRegion)
             
@@ -92,9 +107,10 @@ def record(inputJSON):
             
         
         # Delete the images
-        for fileName in os.listdir('./'):
-            if fileName.endswith(currentImageType):
-                os.remove(fileName)
+        if inputJSON["type"] == "TIMELAPSE":
+            for fileName in os.listdir('./'):
+                if fileName.endswith(currentImageType):
+                    os.remove(fileName)
                 
         # Close Time Lapse
         helper.ClickAndWait(x, y)
